requirements.txt
absl-py==2.0.0
alembic==1.12.0
astor==0.8.1
attrdict==2.0.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
braceexpand==0.1.7
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.3.0
clearml==1.13.1
click==8.1.7
cmaes==0.10.0
colorlog==6.7.0
coqpit==0.0.9
coqui-stt-ctcdecoder==1.4.0
coqui-stt-model-manager==0.0.21
coqui-stt-training==1.4.0
debugpy==1.7.0
decorator==5.1.1
entrypoints==0.4
ffmpeg==1.4
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SocketIO==4.3.2
furl==2.1.3
gast==0.2.2
google-pasta==0.2.0
greenlet==3.0.0
grpcio==1.59.0
h5py==3.8.0
idna==2.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.0.1
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==4.12.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
llvmlite==0.36.0
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
miniaudio==1.59
nest-asyncio==1.5.8
numba==0.53.1
numpy==1.18.5
nvidia-ml-py==12.535.108
opt-einsum==3.3.0
optuna==3.3.0
opuslib==2.0.0
orderedmultidict==1.0.1
packaging==23.2
pandas==1.3.5
parso==0.8.3
pathlib2==2.3.7.post1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
progressbar2==4.2.0
prompt-toolkit==3.0.39
protobuf==3.20.1
psutil==5.9.5
ptyprocess==0.7.0
pycparser==2.21
pydub==0.25.1
Pygments==2.16.1
PyJWT==2.4.0
PyOgg==0.6.14a1
pyparsing==3.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-engineio==3.14.2
python-socketio==4.6.1
python-utils==3.5.2
pytz==2023.3.post1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.25.1
resampy==0.4.2
semver==3.0.2
six==1.16.0
soundfile==0.12.1
soupsieve==2.4.1
sox==1.4.1
SQLAlchemy==2.0.22
stt==1.4.0
tensorboard==1.15.0
tensorflow==1.15.4
tensorflow-estimator==1.15.1
termcolor==2.3.0
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
Unidecode==1.3.7
urllib3==1.26.18
wcwidth==0.2.8
webdataset==0.1.103
webrtcvad==2.0.10
Werkzeug==2.0.3
wrapt==1.15.0
zipp==3.15.0
